apis/v1: Split PodMonitor types in their own file
