+++ Makefile
@@ -19,6 +19,7 @@ VERSION?=$(shell cat VERSION | tr -d " \t\n\r")
 
 TYPES_V1_TARGET := pkg/apis/monitoring/v1/types.go
 TYPES_V1_TARGET += pkg/apis/monitoring/v1/alertmanager_types.go
+TYPES_V1_TARGET += pkg/apis/monitoring/v1/podmonitor_types.go
 TYPES_V1_TARGET += pkg/apis/monitoring/v1/probe_types.go
 TYPES_V1_TARGET += pkg/apis/monitoring/v1/thanos_types.go
