+++ pkg/apis/monitoring/v1/types.go
@@ -36,10 +36,6 @@ const (
 	ServiceMonitorsKind   = "ServiceMonitor"
 	ServiceMonitorName    = "servicemonitors"
 	ServiceMonitorKindKey = "servicemonitor"
-
-	PodMonitorsKind   = "PodMonitor"
-	PodMonitorName    = "podmonitors"
-	PodMonitorKindKey = "podmonitor"
 )
 
 var resourceToKind = map[string]string{
@@ -1315,113 +1311,9 @@ type Endpoint struct {
 	FilterRunning *bool `json:"filterRunning,omitempty"`
 }
 
-// +genclient
-// +k8s:openapi-gen=true
-// +kubebuilder:resource:categories="prometheus-operator",shortName="pmon"
-
-// PodMonitor defines monitoring for a set of pods.
-type PodMonitor struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-	// Specification of desired Pod selection for target discovery by Prometheus.
-	Spec PodMonitorSpec `json:"spec"`
-}
-
-// PodMonitorSpec contains specification parameters for a PodMonitor.
-// +k8s:openapi-gen=true
-type PodMonitorSpec struct {
-	// The label to use to retrieve the job name from.
-	JobLabel string `json:"jobLabel,omitempty"`
-	// PodTargetLabels transfers labels on the Kubernetes Pod onto the target.
-	PodTargetLabels []string `json:"podTargetLabels,omitempty"`
-	// A list of endpoints allowed as part of this PodMonitor.
-	PodMetricsEndpoints []PodMetricsEndpoint `json:"podMetricsEndpoints"`
-	// Selector to select Pod objects.
-	Selector metav1.LabelSelector `json:"selector"`
-	// Selector to select which namespaces the Endpoints objects are discovered from.
-	NamespaceSelector NamespaceSelector `json:"namespaceSelector,omitempty"`
-	// SampleLimit defines per-scrape limit on number of scraped samples that will be accepted.
-	SampleLimit uint64 `json:"sampleLimit,omitempty"`
-	// TargetLimit defines a limit on the number of scraped targets that will be accepted.
-	TargetLimit uint64 `json:"targetLimit,omitempty"`
-	// Per-scrape limit on number of labels that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelLimit uint64 `json:"labelLimit,omitempty"`
-	// Per-scrape limit on length of labels name that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelNameLengthLimit uint64 `json:"labelNameLengthLimit,omitempty"`
-	// Per-scrape limit on length of labels value that will be accepted for a sample.
-	// Only valid in Prometheus versions 2.27.0 and newer.
-	LabelValueLengthLimit uint64 `json:"labelValueLengthLimit,omitempty"`
-	// Attaches node metadata to discovered targets.
-	// Requires Prometheus v2.35.0 and above.
-	AttachMetadata *AttachMetadata `json:"attachMetadata,omitempty"`
-}
-
 type AttachMetadata struct {
 	// When set to true, Prometheus must have permissions to get Nodes.
 	Node bool `json:"node,omitempty"`
-}
-
-// PodMetricsEndpoint defines a scrapeable endpoint of a Kubernetes Pod serving Prometheus metrics.
-// +k8s:openapi-gen=true
-type PodMetricsEndpoint struct {
-	// Name of the pod port this endpoint refers to. Mutually exclusive with targetPort.
-	Port string `json:"port,omitempty"`
-	// Deprecated: Use 'port' instead.
-	TargetPort *intstr.IntOrString `json:"targetPort,omitempty"`
-	// HTTP path to scrape for metrics.
-	// If empty, Prometheus uses the default value (e.g. `/metrics`).
-	Path string `json:"path,omitempty"`
-	// HTTP scheme to use for scraping.
-	Scheme string `json:"scheme,omitempty"`
-	// Optional HTTP URL parameters
-	Params map[string][]string `json:"params,omitempty"`
-	// Interval at which metrics should be scraped
-	// If not specified Prometheus' global scrape interval is used.
-	Interval Duration `json:"interval,omitempty"`
-	// Timeout after which the scrape is ended
-	// If not specified, the Prometheus global scrape interval is used.
-	ScrapeTimeout Duration `json:"scrapeTimeout,omitempty"`
-	// TLS configuration to use when scraping the endpoint.
-	TLSConfig *PodMetricsEndpointTLSConfig `json:"tlsConfig,omitempty"`
-	// Secret to mount to read bearer token for scraping targets. The secret
-	// needs to be in the same namespace as the pod monitor and accessible by
-	// the Prometheus Operator.
-	BearerTokenSecret v1.SecretKeySelector `json:"bearerTokenSecret,omitempty"`
-	// HonorLabels chooses the metric's labels on collisions with target labels.
-	HonorLabels bool `json:"honorLabels,omitempty"`
-	// HonorTimestamps controls whether Prometheus respects the timestamps present in scraped data.
-	HonorTimestamps *bool `json:"honorTimestamps,omitempty"`
-	// BasicAuth allow an endpoint to authenticate over basic authentication.
-	// More info: https://prometheus.io/docs/operating/configuration/#endpoint
-	BasicAuth *BasicAuth `json:"basicAuth,omitempty"`
-	// OAuth2 for the URL. Only valid in Prometheus versions 2.27.0 and newer.
-	OAuth2 *OAuth2 `json:"oauth2,omitempty"`
-	// Authorization section for this endpoint
-	Authorization *SafeAuthorization `json:"authorization,omitempty"`
-	// MetricRelabelConfigs to apply to samples before ingestion.
-	MetricRelabelConfigs []*RelabelConfig `json:"metricRelabelings,omitempty"`
-	// RelabelConfigs to apply to samples before scraping.
-	// Prometheus Operator automatically adds relabelings for a few standard Kubernetes fields.
-	// The original scrape job's name is available via the `__tmp_prometheus_job_name` label.
-	// More info: https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config
-	RelabelConfigs []*RelabelConfig `json:"relabelings,omitempty"`
-	// ProxyURL eg http://proxyserver:2195 Directs scrapes to proxy through this endpoint.
-	ProxyURL *string `json:"proxyUrl,omitempty"`
-	// FollowRedirects configures whether scrape requests follow HTTP 3xx redirects.
-	FollowRedirects *bool `json:"followRedirects,omitempty"`
-	// Whether to enable HTTP2.
-	EnableHttp2 *bool `json:"enableHttp2,omitempty"`
-	// Drop pods that are not running. (Failed, Succeeded). Enabled by default.
-	// More info: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
-	FilterRunning *bool `json:"filterRunning,omitempty"`
-}
-
-// PodMetricsEndpointTLSConfig specifies TLS configuration parameters.
-// +k8s:openapi-gen=true
-type PodMetricsEndpointTLSConfig struct {
-	SafeTLSConfig `json:",inline"`
 }
 
 // OAuth2 allows an endpoint to authenticate with OAuth2.
@@ -1619,17 +1511,6 @@ type ServiceMonitorList struct {
 	Items []*ServiceMonitor `json:"items"`
 }
 
-// PodMonitorList is a list of PodMonitors.
-// +k8s:openapi-gen=true
-type PodMonitorList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard list metadata
-	// More info: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-	// List of PodMonitors
-	Items []*PodMonitor `json:"items"`
-}
-
 // MetadataConfig configures the sending of series metadata to the remote storage.
 // +k8s:openapi-gen=true
 type MetadataConfig struct {
@@ -1692,16 +1573,6 @@ func (l *ServiceMonitor) DeepCopyObject() runtime.Object {
 
 // DeepCopyObject implements the runtime.Object interface.
 func (l *ServiceMonitorList) DeepCopyObject() runtime.Object {
-	return l.DeepCopy()
-}
-
-// DeepCopyObject implements the runtime.Object interface.
-func (l *PodMonitor) DeepCopyObject() runtime.Object {
-	return l.DeepCopy()
-}
-
-// DeepCopyObject implements the runtime.Object interface.
-func (l *PodMonitorList) DeepCopyObject() runtime.Object {
 	return l.DeepCopy()
 }
